requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
openai>=1.0.0
redis>=5.0.0
python-dotenv>=1.0.0
httpx>=0.25.0
pytest>=7.4.0
pydantic>=2.4.0
